Rename darion: fix up references in code
