+++ .gitignore
@@ -78,7 +78,7 @@ checkstyle-result.xml
 /titus-executor
 /root/apps/titus-executor/bin/titus-executor
 /root/apps/titus-executor/bin/titus-reaper
-/root/apps/titus-executor/bin/titus-darion
+/root/apps/titus-executor/bin/titus-logviewer
 /root/apps/titus-executor/bin/titus-metadata-service
 /root/apps/titus-executor/bin/titus-mount
 /root/apps/titus-executor/bin/tini-static
